refactor(workspace): :recycle: rename core to lib
