+++ packages/workspace/src/generators/contributors/generator.ts
@@ -1,6 +1,6 @@
 import { formatFiles, Tree } from '@nrwl/devkit';
 
-import { getGitRepoSlug, readmeFile } from '../core';
+import { getGitRepoSlug, readmeFile } from '../lib';
 
 export async function contributorsGenerator(tree: Tree) {
   if (!tree.exists(readmeFile)) {
