+++ packages/workspace/src/generators/eslint/generator.spec.ts
@@ -1,7 +1,7 @@
 import { addProjectConfiguration, installPackagesTask, Tree } from '@nrwl/devkit';
 import { createTreeWithEmptyWorkspace } from '@nrwl/devkit/testing';
 
-import { eslintConfigFile, lintWorkspaceTask, readEsLintConfig, writeEsLintConfig } from '../core';
+import { eslintConfigFile, lintWorkspaceTask, readEsLintConfig, writeEsLintConfig } from '../lib';
 import { eslintGenerator } from './generator';
 import { eslintSchematic } from './generator.compat';
