+++ packages/workspace/src/generators/lib/add-dependency-to-package-json.spec.ts
