+++ packages/workspace/src/generators/lib/add-dev-dependency-to-package-json.spec.ts
