+++ packages/workspace/src/generators/lib/add-implicit-dependency-to-nx-config.spec.ts
