+++ packages/workspace/src/generators/lib/add-script-to-package-json.ts
