+++ packages/workspace/src/generators/lib/eslint-config.spec.ts
