+++ packages/workspace/src/generators/lib/eslint-config.ts
