+++ packages/workspace/src/generators/lib/get-npm-package-version.ts
