+++ packages/workspace/src/generators/lib/index.ts
@@ -3,7 +3,6 @@ export * from './add-dependency-to-package-json';
 export * from './add-dev-dependency-to-package-json';
 export * from './add-implicit-dependency-to-nx-config';
 export * from './add-script-to-package-json';
-export * from './codecov';
 export * from './eslint-config';
 export * from './format-workspace-task';
 export * from './get-git-repo';
@@ -15,7 +14,6 @@ export * from './nx';
 export * from './package-json';
 export * from './prettier';
 export * from './readme';
-export * from './renovate-config-validator-task';
 export * from './renovate';
 export * from './set';
 export * from './slash';
