+++ packages/workspace/src/generators/lib/join-normalize.spec.ts
