+++ packages/workspace/src/generators/lib/join-normalize.ts
