+++ packages/workspace/src/generators/lib/nx.ts
