+++ packages/workspace/src/generators/lib/package-json.spec.ts
