+++ packages/workspace/src/generators/lib/package-json.ts
