+++ packages/workspace/src/generators/lib/prettier.ts
