+++ packages/workspace/src/generators/lib/readme.ts
