+++ packages/workspace/src/generators/lib/slash.spec.ts
