+++ packages/workspace/src/generators/lib/slash.ts
