+++ packages/workspace/src/generators/lib/tslint-config.ts
