+++ packages/workspace/src/generators/renovate/generator.ts
@@ -1,14 +1,8 @@
 import { formatFiles, generateFiles, Tree } from '@nrwl/devkit';
 import { parseDocument, Scalar, stringify, YAMLSeq } from 'yaml';
 
-import {
-  ciFile,
-  getGitRepoSlug,
-  joinNormalize,
-  renovateBranch,
-  renovateCiFile,
-  renovateConfigValidatorTask,
-} from '../core';
+import { ciFile, getGitRepoSlug, joinNormalize, renovateBranch, renovateCiFile } from '../lib';
+import { renovateConfigValidatorTask } from './renovate-config-validator-task';
 import { RenovateGeneratorSchema } from './schema';
 
 export async function renovateGenerator(tree: Tree, options: RenovateGeneratorSchema) {
