+++ packages/workspace/src/generators/renovate/renovate-config-validator-task.spec.ts
