+++ packages/workspace/src/generators/renovate/renovate-config-validator-task.ts
@@ -2,7 +2,7 @@ import { execSync } from 'child_process';
 
 import { Tree } from '@nrwl/devkit';
 
-import { joinNormalize } from './join-normalize';
+import { joinNormalize } from '../lib';
 
 export function renovateConfigValidatorTask(tree: Tree): boolean {
   try {
